Patch1

+++ src/main/java/com/university/librarymanagementsystem/dto/curriculum/BookReferenceDTO.java
@@ -20,7 +20,7 @@ public class BookReferenceDTO {
     private String course_name;
 
     // @JsonProperty("books_id")
-    private Long book_id;
+    private int book_id;
     private String book_name;
     private String isbn10;
     private String isbn13;
